+++ libs/containers/Message.lua
@@ -403,7 +403,7 @@ function get.mentionedChannels(self)
 		self._mentioned_channels = ArrayIterable(ids, function(id)
 			local guild = client._channel_map[id]
 			if guild then
-				return guild._text_channels:get(id) or guild._voice_channels:get(id)
+				return guild._text_channels:get(id) or guild._voice_channels:get(id) or guild._categories:get(id)
 			else
 				return client._private_channels:get(id) or client._group_channels:get(id)
 			end